prepare on the l9

+++ wedding/src/main/java/cz/engeto/wedding/model/Gift.java
@@ -1,13 +1,12 @@
 package cz.engeto.wedding.model;
 
-import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class Gift {
 
-  private Long id;
-  private String description;
-  private BigDecimal price;
+    private Long id;
+    private String description;
+    private Long price;
 
 }
